// src/features/reset-password/reset-password-schema.ts
import z from 'zod';

export const ResetPasswordSchema = z.object({
    email: z.string().email(),
});

export type ResetPasswordType = z.infer<typeof ResetPasswordSchema>;

export const ResetPasswordResponseSchema = z.object({
    success: z.boolean(),
    name: z.string(),
    message: z.string(),
    httpCode: z.number(),
    data: z.object({}),
});

export type ResetPasswordResponse = z.infer<typeof ResetPasswordResponseSchema>;

export type ResetPasswordError = {
    name: string;
    cause: string;
    hint: string;
    stack: string;
};